microservices/embedding_vllm/pyproject.toml
[project]
name = "embedding-grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.75.1",
    "grpcio-tools>=1.75.1",
    "sentence-transformers>=5.1.1",
    "vllm>=0.11.0",
]
